refactor(config): move satellite configs (flatten clean-up)
