+++ packages/config/src/index.ts
@@ -3,16 +3,16 @@ export * from './console/console.config';
 export * from './module/module.settings';
 export * from './pkg/juno.package';
 export * from './pkg/juno.package.constants';
+export * from './satellite/config';
+export * from './satellite/configs/assertions.config';
+export * from './satellite/configs/authentication.config';
 export * from './satellite/configs/collections';
+export * from './satellite/configs/datastore.config';
+export * from './satellite/configs/emulator.config';
+export * from './satellite/configs/orbiter.config';
 export * from './satellite/configs/rules';
-export * from './satellite/mainnet/config';
-export * from './satellite/mainnet/configs/assertions.config';
-export * from './satellite/mainnet/configs/authentication.config';
-export * from './satellite/mainnet/configs/datastore.config';
-export * from './satellite/mainnet/configs/emulator.config';
-export * from './satellite/mainnet/configs/orbiter.config';
-export * from './satellite/mainnet/configs/satellite.config';
-export * from './satellite/mainnet/juno.config';
+export * from './satellite/configs/satellite.config';
+export * from './satellite/juno.config';
 export * from './shared/feature.config';
 export * from './shared/storage.config';
 export * from './types/cli.config';
