+++ packages/config/src/satellite/configs/assertions.config.ts
