+++ packages/config/src/satellite/configs/authentication.config.ts
