+++ packages/config/src/satellite/configs/emulator.config.ts
