+++ packages/config/src/satellite/juno.config.ts
