+++ packages/config/src/tests/satellite/configs/authentication.config.spec.ts
@@ -2,8 +2,8 @@ import {
   AuthenticationConfigInternetIdentitySchema,
   AuthenticationConfigRulesSchema,
   AuthenticationConfigSchema
-} from '../../../../satellite/mainnet/configs/authentication.config';
-import {mockModuleIdText, mockUserIdText} from '../../../mocks/principal.mock';
+} from '../../../satellite/configs/authentication.config';
+import {mockModuleIdText, mockUserIdText} from '../../mocks/principal.mock';
 
 describe('authentication.config', () => {
   describe('AuthenticationConfigInternetIdentitySchema', () => {
