+++ packages/config/src/tests/satellite/configs/satellite.config.spec.ts
@@ -1,5 +1,5 @@
-import {SatelliteConfigOptionsSchema} from '../../../../satellite/mainnet/configs/satellite.config';
-import {mockModuleIdText} from '../../../mocks/principal.mock';
+import {SatelliteConfigOptionsSchema} from '../../../satellite/configs/satellite.config';
+import {mockModuleIdText} from '../../mocks/principal.mock';
 
 describe('satellite.config', () => {
   describe('SatelliteConfigOptionsSchema', () => {
